fix(target): flaws on SSR and `bottom`, `right` position
Also, extract types from `index.ts` to `types.ts`
