+++ package.json
@@ -80,5 +80,5 @@
   },
   "type": "module",
   "types": "./dist/index.d.ts",
-  "version": "4.2.0"
+  "version": "4.2.0-beta.0"
 }
